+++ src/Event/ApplicationStartup.php
@@ -6,4 +6,8 @@ namespace Yiisoft\Yii\Console\Event;
 
 final class ApplicationStartup
 {
+    public function __construct(
+        public ?string $commandName = null,
+    ) {
+    }
 }